Refactoring the managers
