+++ .env
@@ -24,8 +24,8 @@ APP_SECRET=c46abbb33b3b73c46a808e5560f2d6f1
 # Format described at https://www.doctrine-project.org/projects/doctrine-dbal/en/latest/reference/configuration.html#connecting-using-a-url
 # For an SQLite database, use: "sqlite:///%kernel.project_dir%/var/data.db"
 # Configure your db driver and server_version in config/packages/doctrine.yaml
-DATABASE_URL=mysql://root:ahmedkhaled@127.0.0.1:3306/botmetask
-#DATABASE_URL=mysql://b632eaadb88a3a:85cb8226@us-cdbr-iron-east-02.cleardb.net/heroku_74eb7a8349f5e48?reconnect=true
+#DATABASE_URL=mysql://root:ahmedkhaled@127.0.0.1:3306/botmetask
+DATABASE_URL=mysql://b632eaadb88a3a:85cb8226@us-cdbr-iron-east-02.cleardb.net/heroku_74eb7a8349f5e48?reconnect=true
 ###< doctrine/doctrine-bundle ###
 
 #Facebook Messenger
