+++ .env.example
@@ -24,5 +24,11 @@ APP_SECRET=c46abbb33b3b73c46a808e5560f2d6f1
 # Format described at https://www.doctrine-project.org/projects/doctrine-dbal/en/latest/reference/configuration.html#connecting-using-a-url
 # For an SQLite database, use: "sqlite:///%kernel.project_dir%/var/data.db"
 # Configure your db driver and server_version in config/packages/doctrine.yaml
-DATABASE_URL=mysql://root:root@127.0.0.1:3306/botmetask
+#DATABASE_URL=mysql://root:ahmedkhaled@127.0.0.1:3306/botmetask
+DATABASE_URL=mysql://b632eaadb88a3a:85cb8226@us-cdbr-iron-east-02.cleardb.net/heroku_74eb7a8349f5e48?reconnect=true
 ###< doctrine/doctrine-bundle ###
+
+#Facebook Messenger
+facebook_token=EAATwOmk77nUBAK6zcyO6PWmNvx6g0qvuhx4o1jfpcy1tS4GxSMK9FQSbrl1eqCa8WBaJzMepBQfjjDf18Wy0CbcJPL8goHamnqSJ0z2COoECQIPlUOp0iMxNSIXZCqj8lpP3uxtPdOmo4MORCqXgKqoZAz6HoOnVZAjnVZC9nGZCUu6KgZAxZCo5OLUbfpTVZCAZD
+facebook_app_secret=f470eb62820bcc372c30bfaecfc73439
+facebook_verification=thisisAgoodDay
