+++ config/packages/doctrine.yaml
@@ -24,9 +24,8 @@ doctrine:
                 type: annotation
                 dir: '%kernel.project_dir%/src/Component/Item/Model'
                 prefix: 'App\Component\Item\Model'
-            App:
+            App\Component\User\Model:
                 is_bundle: false
                 type: annotation
-                dir: '%kernel.project_dir%/src/Entity'
-                prefix: 'App\Entity'
-                alias: App
+                dir: '%kernel.project_dir%/src/Component/User/Model'
+                prefix: 'App\Component\User\Model'
