+++ src/Controller/SecurityController.php
@@ -2,9 +2,9 @@
 
 namespace App\Controller;
 
-use App\Entity\User;
+use App\Component\User\Model\User;
 use App\Form\UserType;
-use App\Repository\UserRepository;
+use App\Component\User\Repository\UserRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
